Fixed shears not working under particular settings Fixed potential issue saving JSON files Fixed potential crash with null ore dictionaries Fixed potential crash with null mining agent Updated project files Minor code cleanup

+++ src/main/java/oreexcavation/utils/XPHelper.java
@@ -44,7 +44,7 @@ public class XPHelper
 	
 	public static long getPlayerXP(EntityPlayer player)
 	{
-		return getLevelXP(player.experienceLevel) + (long)Math.round(xpBarCap(player) * (double)player.experience);
+		return getLevelXP(player.experienceLevel) + Math.round(xpBarCap(player) * (double)player.experience);
 	}
 	
 	public static long xpBarCap(EntityPlayer player)
